Do a housekeeping: rename classes and set a feature flag
