+++ pitaya-sharp/NPitaya/src/Metrics/MetricsConfiguration.cs
@@ -2,12 +2,14 @@ namespace NPitaya.Metrics
 {
     public class MetricsConfiguration
     {
+        internal readonly bool IsEnabled;
         internal readonly string Host;
         internal readonly string Port;
         internal readonly string Namespace;
 
-        public MetricsConfiguration(string host, string port, string ns)
+        public MetricsConfiguration(bool isEnabled, string host, string port, string ns)
         {
+            IsEnabled = isEnabled;
             Host = host;
             Port = port;
             Namespace = ns;
